+++ Tester/MasterViewController.cs
@@ -110,12 +110,12 @@ namespace Tester
         get { return objects; }
       }
       // Customize the number of sections in the table view.
-      public override int NumberOfSections(UITableView tableView)
+      public override nint NumberOfSections(UITableView tableView)
       {
         return 1;
       }
 
-      public override int RowsInSection(UITableView tableview, int section)
+      public override nint RowsInSection(UITableView tableview, nint section)
       {
         return objects.Count;
       }
